new test cases for UserServiceImpl modify UserRepository, from LONG to Integer

+++ OrderService/src/main/java/com/suraev/repository/UserEntityRepository.java
@@ -3,5 +3,5 @@ package com.suraev.repository;
 import com.suraev.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserEntityRepository extends JpaRepository<User, Long> {
+public interface UserEntityRepository extends JpaRepository<User, Integer> {
 }
